fix press page build and disable unescaped lint
